etl/requirements.txt
pydantic==2.4.2
pydantic-settings==2.0.3
kafka-python==2.0.2
clickhouse-driver==0.2.6
aiokafka==0.8.1
